requirements.txt
feedparser==6.0.8
Flask==2.0.3
flask_cors==3.0.10
hdbscan==0.8.28
numpy==1.22.2
python-dotenv==0.19.2
python_dateutil==2.8.2
transformers==4.16.2
umap-learn==0.5.2
torch